Move Chef::OscUser back to Chef::User namespace and new user code to Chef::UserV1.
Also, have Chef::User (formally Chef::OscUser) use API V0 to make requests.
