+++ spec/unit/knife/osc_user_list_spec.rb
@@ -30,7 +30,7 @@ describe Chef::Knife::OscUserList do
   end
 
   it 'lists the users' do
-    expect(Chef::OscUser).to receive(:list)
+    expect(Chef::User).to receive(:list)
     expect(@knife).to receive(:format_list_for_display)
     @knife.run
   end
