+++ spec/unit/knife/user_edit_spec.rb
@@ -36,7 +36,7 @@ describe Chef::Knife::UserEdit do
   context "when the username field is not supported by the server" do
     before do
       allow(knife).to receive(:run_osc_11_user_edit).and_raise(SystemExit)
-      allow(Chef::User).to receive(:load).and_return({"username" => nil})
+      allow(Chef::UserV1).to receive(:load).and_return({"username" => nil})
     end
 
     it "displays the osc warning" do
@@ -52,7 +52,7 @@ describe Chef::Knife::UserEdit do
 
   it 'loads and edits the user' do
     data = { "username" => "my_user" }
-    allow(Chef::User).to receive(:load).with("my_user").and_return(data)
+    allow(Chef::UserV1).to receive(:load).with("my_user").and_return(data)
     expect(knife).to receive(:edit_data).with(data).and_return(data)
     knife.run
   end
